+++ langchain-core/src/runnables/tests/runnable_stream_log.test.ts
@@ -13,9 +13,13 @@ import { RunLog } from "../../tracers/log_stream.js";
 import { RunnableSequence, RunnableMap } from "../base.js";
 import {
   FakeLLM,
+  FakeStreamingLLM,
   FakeChatModel,
   FakeRetriever,
 } from "../../utils/testing/index.js";
+import { SystemMessage, HumanMessage } from "../../messages/index.js";
+import { CommaSeparatedListOutputParser } from "../../output_parsers/list.js";
+import { ChatPromptValue } from "../../prompt_values.js";
 
 test("Runnable streamLog method", async () => {
   const promptTemplate = PromptTemplate.fromTemplate("{input}");
@@ -93,4 +97,213 @@ test("Runnable streamLog method with a more complicated sequence", async () => {
   expect(
     (finalState as RunLog).state.logs.CUSTOM_NAME.final_output.output
   ).toEqual(JSON.stringify(retrieverOutputDocs));
+});
+
+test.only("Test stream log aggregation", async () => {
+  const prompt = ChatPromptTemplate.fromMessages([
+    ["system", "You are a nice assistant"],
+    ["human", "{question}"],
+  ]);
+  const llm = new FakeStreamingLLM({
+    responses: ["tomato, lettuce, onion"],
+  });
+  const parser = new CommaSeparatedListOutputParser({});
+  const chain = prompt.pipe(llm).pipe(parser);
+  const logStream = await chain.streamLog({
+    question: "what is up?",
+  });
+  const chunks = [];
+  for await (const chunk of logStream) {
+    chunks.push(chunk);
+  }
+  expect(chunks).toMatchObject([
+    {
+      ops: [
+        {
+          op: "replace",
+          path: "",
+          value: {
+            id: expect.any(String),
+            streamed_output: [],
+            logs: {},
+          },
+        },
+      ],
+    },
+    {
+      ops: [
+        {
+          op: "add",
+          path: "/logs/ChatPromptTemplate",
+          value: {
+            id: expect.any(String),
+            name: "ChatPromptTemplate",
+            type: "prompt",
+            tags: ["seq:step:1"],
+            metadata: {},
+            start_time: expect.any(String),
+            streamed_output: [],
+            streamed_output_str: [],
+          },
+        },
+      ],
+    },
+    {
+      ops: [
+        {
+          op: "add",
+          path: "/logs/ChatPromptTemplate/final_output",
+          value: new ChatPromptValue([
+            new SystemMessage("You are a nice assistant"),
+            new HumanMessage("what is up?"),
+          ]),
+        },
+        {
+          op: "add",
+          path: "/logs/ChatPromptTemplate/end_time",
+          value: expect.any(String),
+        },
+      ],
+    },
+    {
+      ops: [
+        {
+          op: "add",
+          path: "/logs/FakeStreamingLLM",
+          value: {
+            id: expect.any(String),
+            name: "FakeStreamingLLM",
+            type: "llm",
+            tags: ["seq:step:2"],
+            metadata: {},
+            start_time: expect.any(String),
+            streamed_output: [],
+            streamed_output_str: [],
+          },
+        },
+      ],
+    },
+    {
+      ops: [
+        {
+          op: "add",
+          path: "/logs/CommaSeparatedListOutputParser",
+          value: {
+            id: expect.any(String),
+            name: "CommaSeparatedListOutputParser",
+            type: "parser",
+            tags: ["seq:step:3"],
+            metadata: {},
+            start_time: expect.any(String),
+            streamed_output: [],
+            streamed_output_str: [],
+          },
+        },
+      ],
+    },
+    {
+      ops: [
+        {
+          op: "add",
+          path: "/logs/CommaSeparatedListOutputParser/streamed_output/-",
+          value: ["tomato"],
+        },
+      ],
+    },
+    {
+      ops: [
+        {
+          op: "add",
+          path: "/streamed_output/-",
+          value: ["tomato"],
+        },
+      ],
+    },
+    {
+      ops: [
+        {
+          op: "add",
+          path: "/logs/CommaSeparatedListOutputParser/streamed_output/-",
+          value: ["lettuce"],
+        },
+      ],
+    },
+    {
+      ops: [
+        {
+          op: "add",
+          path: "/streamed_output/-",
+          value: ["lettuce"],
+        },
+      ],
+    },
+    {
+      ops: [
+        {
+          op: "add",
+          path: "/logs/FakeStreamingLLM/final_output",
+          value: {
+            generations: [
+              [
+                {
+                  text: "tomato, lettuce, onion",
+                  generationInfo: {},
+                },
+              ],
+            ],
+          },
+        },
+        {
+          op: "add",
+          path: "/logs/FakeStreamingLLM/end_time",
+          value: expect.any(String),
+        },
+      ],
+    },
+    {
+      ops: [
+        {
+          op: "add",
+          path: "/logs/CommaSeparatedListOutputParser/streamed_output/-",
+          value: ["onion"],
+        },
+      ],
+    },
+    {
+      ops: [
+        {
+          op: "add",
+          path: "/streamed_output/-",
+          value: ["onion"],
+        },
+      ],
+    },
+    {
+      ops: [
+        {
+          op: "add",
+          path: "/logs/CommaSeparatedListOutputParser/final_output",
+          value: {
+            output: ["tomato", "lettuce", "onion"],
+          },
+        },
+        {
+          op: "add",
+          path: "/logs/CommaSeparatedListOutputParser/end_time",
+          value: expect.any(String),
+        },
+      ],
+    },
+    {
+      ops: [
+        {
+          op: "replace",
+          path: "/final_output",
+          value: {
+            output: ["tomato", "lettuce", "onion"],
+          },
+        },
+      ],
+    },
+  ]);
 });
